requirements-linux.txt
opengate==10.0.1 # see TODOs.md for 10.0.2
awkward-pandas
napari[all]==0.5.5 # 0.5.6 not compatible with latest napari-bbox 0.0.9
napari-bbox==0.0.9
# TODO: is pyvista also not installed with opengate, as with MacOS?